+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/add_service/add_service.go
@@ -332,7 +332,8 @@ func ValidateAndConvertConfigAndReadyCondition(
 		locatorOfModuleInWhichThisBuiltInIsBeingCalled,
 		packageId,
 		packageContentProvider,
-		packageReplaceOptions, imageDownloadMode)
+		packageReplaceOptions,
+		imageDownloadMode)
 	if interpretationErr != nil {
 		return nil, nil, interpretationErr
 	}